Quick addition.
Added 500 random users so that we have more options of people to like/dislike
Added a parameter to the cat api so that it fetches small jags.
fixed a login bug where user won't be notified if they typed the wrong password.
I also removed UserConnection.sync() apparently you only need to do that once to create the table.